% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_syn.R
\name{check_syn}
\alias{check_syn}
\title{Function to identify synonyms of species' name that is absent in a database}
\usage{
check_syn(myvec, dbvec, db.syn = "itis")
}
\arguments{
\item{myvec}{vector of species' name that is compared to dbvec}

\item{dbvec}{vector of species' name available in a database}

\item{db.syn}{database to query to find synonyms, either itis, tropicos, nbn, worms, or pow (see taxize::synonyms)}
}
\value{
A dataframe with one column indicating the missing species' name in the database and one colum indicating its potential synonyms
}
\description{
Function to identify synonyms of species' name that is absent in a database
}
